namespace adjusts for validation
