+++ tests/integration/Validation.XmlValidator.test.js
@@ -22,45 +22,46 @@ const { describe, it } = require('mocha')
 
 const { escapeRegExp } = require('../_helpers/stringFunctions')
 
-let hasAjv = true
+let hasLibXml = true
 try {
-  require('ajv')
+  require('libxmljs2')
 } catch {
-  hasAjv = false
+  hasLibXml = false
 }
 
 const {
   Spec: { Version },
   Validation: {
     ValidationError, NotImplementedError, MissingOptionalDependencyError,
-    Validators: { JsonStrictValidator }
+    XmlValidator
   }
 } = require('../../')
 
-describe('Validation.Validators.JsonStrictValidator', () => {
+describe('Validation.XmlValidator', () => {
   [
-    Version.v1dot0,
-    Version.v1dot1
+    // none so far
   ].forEach((version) => describe(version, () => {
     it('throws not implemented', async () => {
-      const validator = new JsonStrictValidator(version)
+      const validator = new XmlValidator(version)
       return assert.rejects(
-        () => validator.validate('{}'),
+        () => validator.validate('<bom/>'),
         (err) => err instanceof NotImplementedError
       )
     })
   }));
 
   [
+    Version.v1dot0,
+    Version.v1dot1,
     Version.v1dot2,
     Version.v1dot3,
     Version.v1dot4
   ].forEach((version) => describe(version, () => {
-    if (!hasAjv) {
+    if (!hasLibXml) {
       it('throws MissingOptionalDependencyError', async () => {
-        const validator = new JsonStrictValidator(version)
+        const validator = new XmlValidator(version)
         await assert.rejects(
-          () => validator.validate('{}'),
+          () => validator.validate('<bom/>'),
           (err) => err instanceof MissingOptionalDependencyError
         )
       })
@@ -68,16 +69,17 @@ describe('Validation.Validators.JsonStrictValidator', () => {
     }
 
     it('invalid throws', async () => {
-      const validator = new JsonStrictValidator(version)
-      const input = JSON.stringify({
-        bomFormat: 'CycloneDX',
-        specVersion: version,
-        components: [{
-          type: 'library',
-          name: 'bar',
-          unknown: 'undefined' // << undefined/additional property
-        }]
-      })
+      const validator = new XmlValidator(version)
+      const input = `<?xml version="1.0" encoding="UTF-8"?>
+        <bom xmlns="http://cyclonedx.org/schema/bom/${version}">
+          <components>
+            <component type="library">
+              <name>bar</name>
+              <version>1.337</version>
+              <unknown>undefined</unknown><!-- << undefined/additional property -->
+            </component>
+          </components>
+        </bom>`
       return assert.rejects(
         () => validator.validate(input),
         (err) => {
@@ -90,16 +92,17 @@ describe('Validation.Validators.JsonStrictValidator', () => {
     })
 
     it('valid passes', async () => {
-      const validator = new JsonStrictValidator(version)
-      const input = JSON.stringify({
-        bomFormat: 'CycloneDX',
-        specVersion: version,
-        components: [{
-          type: 'library',
-          name: 'foo',
-          version: '1.337'
-        }]
-      })
+      const validator = new XmlValidator(version)
+      const input = `<?xml version="1.0" encoding="UTF-8"?>
+        <bom xmlns="http://cyclonedx.org/schema/bom/${version}">
+          <components>
+            <component type="library">
+              <name>bar</name>
+              <version>1.337</version>
+              ${version === '1.0' ? '<modified>false</modified>' : ''}
+            </component>
+          </components>
+        </bom>`
       await validator.validate(input)
     })
   }))
